chore: add workflows
Adds reusable workflows for running tests, running go vet, and running
go fmt on changed files, and a pull request workflow that runs them all.
The individual steps are reusable workflows instead of inline in the
pull request workflow so the test and vet steps can be reused in an
on-push workflow to create issues if there are somehow broken tests or
vet issues in main without without worrying about the fmt step.

There's also a get-go-version workflow that injects the go version from
the go mod into the workflows so they're always in sync.