feat: type registry (#1)
Creates the Registry type with generic functions for registering
implementation types. Each registration includes a target type (the
type we're registering an implementation for) and an implementation type
(the concrete type used to implement the requested type). Some types
will have default factories in which case it suffices to register an
implementation type for a target type using RegisterType. The
RegisterFactory function allows the registration of types that don't
have default factories as well as custom initialization logic for types
that do have defaults. NOTE that resolution is not implemented yet and
the registration semantics reflect the INTENDED behaviour.

Also populates the README with concepts and intentions for
yet-to-be-implemented types such as Resolver.