chore: express memory model consistently
The RootProvider and Scope types hold pointers to their registrations
and shared value pools meaning that copies of these types are views ino
the same shared data. This ensures there is no need to use pointers to
these types. The previous implementation had the correct behaviour but
used pointers where they were not necessary.

- The Scope type held a pointer to a RootProvider instead of a
  RootProvider.

- The NewScope method on the RootProvider type took a pointer receiver
  instead of a value receiver.

No testing has been done to determine the use of pointers for method
receivers and/or the Scope type holding pointers to the RootProvider
have meaningful performance implications -- this change is being made
only for the sake of consistency with the intention that these types can
be used without pointers.