chore: update README and package doc
Updates the README to use the term "unit of work" instead of "request"
when describing Scopes and replaces the term "request-scoped" with just
plain "scoped". Applications that handle requests are a good example of
a time when Scopes are useful, but not every case for using Scopes is
well described by the term "request".

Normalizes the package doc comment and the README intro.